% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/searchDB.R
\name{getAllCells}
\alias{getAllCells}
\title{Obtain all cell lines tested on a set of drugs}
\usage{
getAllCells(drug)
}
\arguments{
\item{drug}{[vector] vector of drugs to search}
}
\value{
list of data.frame or NA containing search results
}
\description{
Same as title
}
\examples{
getAllCells(c("Erlotinib", "17-AAG"))
}

