% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/searchDB.R
\name{getAllDrugs}
\alias{getAllDrugs}
\title{Obtain all drugs tested on a set of cell lines}
\usage{
getAllDrugs(cell)
}
\arguments{
\item{cell}{[vector] vector of cell lines to search}
}
\value{
list of data.frame or NA containing search results
}
\description{
Same as title
}
\examples{
getAllDrugs(c("MCF7", "1321N1"))
}

