% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/getCuration.R
\name{getCuration}
\alias{getCuration}
\title{Obtain a curation}
\usage{
getCuration(curation = c("cell", "drug", "tissue", "histology",
  "gene_expression"), study = NULL, subtype = NULL)
}
\arguments{
\item{curation}{the curation to be requested:

"cell" is listing of cell lines in each study and corresponding synonyms

"drug" is listing of drugs in each study and corresponding synonyms
"tissue" is listing of cell lines in each study, their corresponding tissue types,
and matching of tissue types in each study to COSMIC tissue types

"histology" is listing of cell lines in each study, their corresponding histological subtypes,
 matching between subtypes based on manual curation by BHK lab

"gene_expression" is listing of gene expression metadata by study}

\item{study}{optional vector of studies to subset curation
'}

\item{subtype}{optional vector of subtype so only curation for that subtype is displayed
       only valid for "tissue", "histology" and "gene expression" curations}
}
\value{
list containing curations if found or NA if curation not found
}
\description{
\code{getCuration} requests a particular curation
 according to user specified parameters from PharmacoDB web service.
}
\examples{
getCuration("cell")
## get the complete cell line curation
getCuration("drug", study=c("CGP", "CCLE"))
## get all drugs in CGP and in CCLE
getCuration("tissue", study=c("CCLE", "GRAY", "GNE"), subtype=c("breast", "testis"))
## get all breast and testis cancer cell lines tested in CCLE and GRAY
getCuration("histology", subtype = "carcinoma")
## get all cell lines which are carcinomas
getCuration("gene_expression", "MEIS")
## returns warning since "MEIS" study does not exist
}

