% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/getData.R
\name{getDataset}
\alias{getDataset}
\title{Obtain a data set according to a study name}
\usage{
getDataset(study, summary = FALSE, stats = NULL)
}
\arguments{
\item{study}{[vector] vector of studies from which to obtain data}

\item{summary}{[boolean] optional parameter of whether or not to return summary data
if summary=TRUE, summary data is returned
if summary=FALSE, dose response data is returned
default is summary=FALSE}

\item{stats}{[vector] vector of summary statistics values to return
default is NULL because default for summary is FALSE}
}
\value{
list of data.frame containing data if found or NA if data does not exist
}
\description{
\code{getDataset} requests a particular dataset
according to user specified study
}
\examples{
getDataset("CCLE") ## get the CCLE dose-response data
getDataset(c("CGP", "GNE")) ## get CGP and GNE dose-response data
getDataset("CTRP", summary=TRUE) ## get all summary statistics for CTRP data
getDataset("CGP", stats = c("IC50_published")) ## get IC50 values in CGP
}

