% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/searchDB.R
\name{getDataStudies}
\alias{getDataStudies}
\title{Get list of studies with dose-response data/summary data in the database}
\usage{
getDataStudies()
}
\value{
vector of studies
}
\description{
Same as title
}

