% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/getData.R
\name{getExpData}
\alias{getExpData}
\title{Obtain a data set according to a study name}
\usage{
getExpData(cellline = NULL, drug = NULL, summary = FALSE, stats = NULL)
}
\arguments{
\item{cellline}{[vector] vector of cell lines for which to obtain data
if cellline=NULL and drug != NULL, then data for all celllines
tested on that vector of drugs is returned.}

\item{drug}{[vector] vector of drugs for which to obtain data
if drug=NULL and cellline != NULL, then data for all
drugs tested on the cellline vector returned

if drug!=NULL and cellline !=NULL, then data for all
experiments in the set cellline, drug pairs for which
there is data is returned}

\item{summary}{[boolean] whether or not to return summary data
if summary=TRUE, summary data is returned
if summary=FALSE, dose response data is returned
default is summary=FALSE}

\item{stats}{[vector] vector of summary statistics values to return
default is stats=NULL because summary=FALSE}
}
\value{
list of data.frame containing data if found or NA if data does not exist
}
\description{
\code{getExpData} requests a particular dataset
according to user specified study
}
\examples{
getExpData(cellline="HCC70") ## get all dose-response curves tested on HCC70
## Get the published IC50 values for all cell lines tested on Erlotinib or 17-AAG
getExpData(drug=c("Erlotinib", "17-AAG"), stats = "IC50_Published"))
## Get all summary statistics for experiments tested on MCF7 and 1321N1
## and with erlotinib or 17-AAG
getExpData(cellline = c("MCF7", "1321N1"), drug = c("erlotinib", "17-AAG"), summary = TRUE)
}

