% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/searchDB.R
\name{listSummaryStats}
\alias{listSummaryStats}
\title{Get list of summary statistics available in a study}
\usage{
listSummaryStats(study)
}
\value{
list of vector of summary statistics
}
\description{
Same as title
}
\examples{
listSummaryStats(c("CCLE", "CGP"))
}

